Final Project: Part 2 Data
Uploading example data set for interactions volume.  Used SQL/Looker to pull the data from the warehouse and did some initial wrangling in Excel in order to obfuscate the data and render data far from company actuals.  The hope here is to practice modeling techniques on this dummy data set and take the learnings back to the company using real data later.